refactor home - Change directory (move Home to home dir) - Refactor Home (Created individual components)
